+++ internal/core/data/application/event_test.go
@@ -35,7 +35,7 @@ const (
 	testUUIDString         = "ca93c8fa-9919-4ec5-85d3-f81b2b6a7bc1"
 	testOriginTime         = 1600666185705354000
 	nonexistentEventID     = "8ad33474-fbc5-11ea-adc1-0242ac120002"
-	testEventCount         = uint32(7778)
+	testEventCount         = int64(7778)
 )
 
 var persistedEvent = models.Event{
@@ -411,9 +411,9 @@ func TestEventsByTimeRange(t *testing.T) {
 
 	dic := mocks.NewMockDIC()
 	dbClientMock := &dbMock.DBClient{}
-	dbClientMock.On("EventCountByTimeRange", event1.Origin, event5.Origin).Return(uint32(5), nil)
+	dbClientMock.On("EventCountByTimeRange", event1.Origin, event5.Origin).Return(int64(5), nil)
 	dbClientMock.On("EventsByTimeRange", event1.Origin, event5.Origin, 0, 10).Return([]models.Event{event5, event4, event3, event2, event1}, nil)
-	dbClientMock.On("EventCountByTimeRange", event2.Origin, event4.Origin).Return(uint32(3), nil)
+	dbClientMock.On("EventCountByTimeRange", event2.Origin, event4.Origin).Return(int64(3), nil)
 	dbClientMock.On("EventsByTimeRange", event2.Origin, event4.Origin, 0, 10).Return([]models.Event{event4, event3, event2}, nil)
 	dbClientMock.On("EventsByTimeRange", event2.Origin, event4.Origin, 1, 2).Return([]models.Event{event3, event2}, nil)
 	dbClientMock.On("EventsByTimeRange", event2.Origin, event4.Origin, 4, 2).Return(nil, errors.NewCommonEdgeX(errors.KindRangeNotSatisfiable, "query objects bounds out of range", nil))
@@ -432,13 +432,13 @@ func TestEventsByTimeRange(t *testing.T) {
 		errorExpected      bool
 		ExpectedErrKind    errors.ErrKind
 		expectedCount      int
-		expectedTotalCount uint32
+		expectedTotalCount int64
 		expectedStatusCode int
 	}{
-		{"Valid - all events", event1.Origin, event5.Origin, 0, 10, false, "", 5, uint32(5), http.StatusOK},
-		{"Valid - events trimmed by latest and oldest", event2.Origin, event4.Origin, 0, 10, false, "", 3, uint32(3), http.StatusOK},
-		{"Valid - events trimmed by latest and oldest and skipped first", event2.Origin, event4.Origin, 1, 2, false, "", 2, uint32(3), http.StatusOK},
-		{"Invalid - bounds out of range", event2.Origin, event4.Origin, 4, 2, true, errors.KindRangeNotSatisfiable, 0, uint32(0), http.StatusRequestedRangeNotSatisfiable},
+		{"Valid - all events", event1.Origin, event5.Origin, 0, 10, false, "", 5, int64(5), http.StatusOK},
+		{"Valid - events trimmed by latest and oldest", event2.Origin, event4.Origin, 0, 10, false, "", 3, int64(3), http.StatusOK},
+		{"Valid - events trimmed by latest and oldest and skipped first", event2.Origin, event4.Origin, 1, 2, false, "", 2, int64(3), http.StatusOK},
+		{"Invalid - bounds out of range", event2.Origin, event4.Origin, 4, 2, true, errors.KindRangeNotSatisfiable, 0, int64(0), http.StatusRequestedRangeNotSatisfiable},
 	}
 	for _, testCase := range tests {
 		t.Run(testCase.name, func(t *testing.T) {
